Back to int32

+++ make_test_snap.py
@@ -121,7 +121,7 @@ def make_ics(filepath, cdim, gdim, boxsize, doner_path):
         # Write the cells metadata
         cell_meta = cell_struct.create_group("Meta-data")
         cell_meta.attrs["dimension"] = np.array(
-            [cdim, cdim, cdim], dtype=np.int64
+            [cdim, cdim, cdim], dtype=np.int32
         )
         cell_meta.attrs["size"] = np.array([cell_size, cell_size, cell_size])
